website_flask/requirements.txt
Flask==2.1.0
pyodbc==4.0.32
azure-core==1.26.4
azure-identity==1.13.0
azure-storage-blob==12.16.0
